+++ app/src/main/java/io/golos/cyber_android/ui/common/formatters/size/SizeFormatter.kt
@@ -1,5 +1,5 @@
 package io.golos.cyber_android.ui.common.formatters.size
 
 interface SizeFormatter {
-    fun format(size: Long): String
+    fun format(size: Int): String
 }
