+++ app/src/main/java/io/golos/cyber_android/ui/screens/subscriptions/Community.kt
@@ -1,3 +1,3 @@
 package io.golos.cyber_android.ui.screens.subscriptions
 
-data class Community (val communityId: String, val name: String, val logo: String?, val followersCount: Long, var isSubscribed: Boolean)
+data class Community (val communityId: String, val name: String, val logo: String?, val followersCount: Int, var isSubscribed: Boolean)
