+++ domain/src/main/java/io/golos/domain/commun_entities/PostDiscussionRaw.kt
@@ -20,7 +20,7 @@ data class PostDiscussionRaw(
 
     val author: DiscussionAuthor,
 
-    val childCount: Long,
+    val childCount: Int,
 
     val shareUrl: String?
 )
